bot-system/pyproject.toml
[tool.poetry]
name = "self-aware-assistant"
version = "0.1.0"
description = "Advanced self-aware personal assistant bot with self-healing capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
loguru = "^0.7.2"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
requests = "^2.31.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"
typer = "^0.9.0"
rich = "^13.7.0"
transformers = "^4.36.0"
torch = "^2.1.2"
numpy = "^1.26.2"
nltk = "^3.8.1"
scikit-learn = "^1.4.0"
supabase = "^2.3.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"

[tool.poetry.group.integrations]
optional = true

[tool.poetry.group.integrations.dependencies]
google-api-python-client = "^2.111.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
slack-sdk = "^3.26.1"
discord-py = "^2.3.2"
boto3 = "^1.34.21"
twilio = "^8.11.1"
sendgrid = "^6.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
